fixes #396 illumos doesn't build (missing NNG_PLATFORM_POSIX ON) fixes #397 Need to cast zoneid fixes #395 sun is predefined on illumos/Solaris fixes #394 alloca needs to #include <alloca.h> fixes #399 Cannot use SVR4.2 specific msghdr fixes #402 getpeerucred needs a NULL initialized ucred fixes #403 syntax error in posix_tcp - attempt to return void fixes #407 illumos getegid wrong fixes #406 nni_idhash_count is dead code fixes #404 idhash typedef redeclared fixes #405 warning: newline not last character in file
This is basically a slew of related bug fixes required to make this
work on illumos.  Note that the fixes are not "complete", because
more work is required to support port events given that epoll is busted
on illumos.

We also fixed a bunch of things that aren't actually "bugs" per se, but
really just warnings.  Silencing them makes things better for everyone.
Apparently not all compilers are equally happy with redundant (but
otherwise identical) typedefs; we use structs in some places instead of
shorter type names to silence these complaints.

Note that IPC permissions (the mode bits on the socket vnode) are not
validated on SunOS systems.  This change includes documentation to reflect
that.

+++ tests/ipc.c
@@ -1,6 +1,6 @@
 //
-// Copyright 2017 Garrett D'Amore <garrett@damore.org>
-// Copyright 2017 Capitar IT Group BV <info@capitar.com>
+// Copyright 2018 Staysail Systems, Inc. <info@staysail.tech>
+// Copyright 2018 Capitar IT Group BV <info@capitar.com>
 //
 // This software is supplied under the terms of the MIT License, a
 // copy of which should be located in the distribution where this
@@ -69,7 +69,7 @@ check_props(nng_msg *msg)
 
 #ifdef NNG_HAVE_GETPEERUCRED
 	So(nng_pipe_getopt_uint64(p, NNG_OPT_IPC_PEER_ZONEID, &id) == 0);
-	So(id == getzoneid());
+	So(id == (uint64_t) getzoneid());
 #else
 	So(nng_pipe_getopt_uint64(p, NNG_OPT_IPC_PEER_ZONEID, &id) ==
 	    NNG_ENOTSUP);
